pudu: C++ Tools for Cleaning Strings

Provides function declarations and inline function definitions that facilitate cleaning strings in C++ code before passing them to R.

See also

Author

Maintainer: Mauricio Vargas Sepulveda m.sepulveda@mail.utoronto.ca (ORCID)

Other contributors:

Munk School of Global Affairs and Public Policy [funder]